rearrange dim and fact generation functions and test suite. create sales fact generator
