--- src/transform/dim_payment_type.py
@@ -1,3 +1,0 @@
-def payment_type_dim(currency_df):
-    currency_df = currency_df.drop(['created_at', 'last_updated'], axis=1)
-    return currency_df
